overwrite fixed

+++ vm/operations/11_sti.c
@@ -13,8 +13,8 @@ void	sti_op(t_general *gen, t_process *process, int op_num, uint32_t *args)
 		|| (args[1] == T_REG && !(args_val[1] >= 1 && args_val[1] <= 16))
 		|| (args[2] == T_REG && !(args_val[2] >= 1 && args_val[2] <= 16)))
 			return ;
-	first = ((args[1] == T_REG) ? process->reg[args_val[1]] : (int)args_val[1]);
-	second = ((args[2] == T_REG) ? process->reg[args_val[2]] : (int)args_val[2]);
+	first = ((args[1] == T_REG) ? process->reg[args_val[1]] : (short)args_val[1]);
+	second = ((args[2] == T_REG) ? process->reg[args_val[2]] : (short)args_val[2]);
 	sum_go = (first + second) % IDX_MOD;
 	sum_go = check_pc(process->pc + sum_go);
 	if (gen->debug)
